+++ test/viking/model/instance_properties/paramRoot.js
@@ -7,5 +7,13 @@
     
         equal(model.paramRoot(), 'model');
     });
+    
+    test("instance.paramRoot returns underscored baseModel.modelName when used as STI", function() {
+        var Boat = Viking.Model.extend('boat');
+        var Carrier = Boat.extend('boat');
+        var model = new Carrier();
+    
+        equal(model.paramRoot(), 'boat');
+    });
 
 }());